+++ docs/src/conf.py
@@ -8,7 +8,7 @@ project = "ska-dlm-client"
 html_title = "SKA DLM Client"
 copyright = "2024, ICRAR"
 author = "Mark Boulton <mark.boulton@uwa.edu.au>"
-release = "1.2.4"
+release = "2.0.0"
 
 extensions = [
     "sphinx.ext.autodoc",