+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Version History
 
-## Planned 1.2.5
+## Development
+
+* The ConfigDB watcher now watches DataProductPersist Flows and checks the status of the related DataProduct Flow.
+
+## 2.0.0
 
 ### Updated